+++ lib/src/effect/EasingMotion.dart
@@ -30,16 +30,16 @@ class EasingMotion extends Motion {
    * + [end] is invoked when the motion ends.
    */
   EasingMotion(this.action, [EasingFunction easing, int period = 500, 
-    int repeat = 1, MotionStart start, MotionEnd end, bool autorun = true]) : 
+    int repeat = 1, MotionStart start, MotionEnd end]) : 
     this.easing = easing, this.period = period, this.repeat = repeat, 
-    duration = repeat * period, super(start, null, end, autorun);
+    duration = repeat * period, super(start, null, end);
   
   /** Construct an EasingMotion by joining multiple EasingMotion.
    */
   EasingMotion.join(List<EasingMotion> motions, [EasingFunction easing, 
-  int period = 500, int repeat = 1, MotionStart start, MotionEnd end, bool autorun = true]) :
+  int period = 500, int repeat = 1, MotionStart start, MotionEnd end]) :
   this(_jointAction(motions), easing: easing, period: period, repeat: repeat,
-  start: _jointStart(motions, start), end: _jointEnd(motions, end), autorun: autorun);
+  start: _jointStart(motions, start), end: _jointEnd(motions, end));
   
   static MotionAction _jointAction(List<EasingMotion> motions) {
     return (num x, MotionState state) {