+++ lib/src/effect/GlowEffect.dart
@@ -22,7 +22,7 @@ class GlowEffect extends EasingMotion {
    */
   GlowEffect(Element element, [int period = 1000, num tempo = 0.3, 
   Color color, int blur = 10, int spread = 2, 
-  MotionStart start, MotionEnd end, bool autorun = true]) : 
+  MotionStart start, MotionEnd end]) : 
   this.element = element, 
   super(null, start: (MotionState state) {
     if (start != null)
@@ -34,7 +34,7 @@ class GlowEffect extends EasingMotion {
     if (end != null)
       end(state);
     
-  }, period: period, autorun: autorun, 
+  }, period: period, 
   easing: (num t) => t < tempo ? (t * t / tempo / tempo) : (1 - t) / (1 - tempo)) {
     
     final int r = color == null ? 255 : color.red.toInt();